min fix

+++ Assets/Internal/Codebase/SaveSystem/SaveData.cs
@@ -8,9 +8,9 @@ namespace Codebase
     public class SaveData
     {
         public int Balance;
-        public int FoodStat;
-        public int MentalStat;
-        public int HealthStat;
-        public int EnergyStat;
+        public float FoodStat;
+        public float MentalStat;
+        public float HealthStat;
+        public float EnergyStat;
     }
 }
